+++ src/test/fixtures/testSTMCubeMakefile.ts
@@ -108,6 +108,7 @@ export const testMakefileInfo: MakeInfo = {
   cxxFlags: [],
   optimization: 'Og',
   customMakefileRules: undefined,
+  makeFlags: [],
 };
 
 export const stm32ForVSCodeResult = `##########################################################################################################################